ai-sales-bot/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
python-multipart==0.0.12
httpx==0.27.0
anthropic==0.37.0
openai==1.50.0
pandas==2.2.0
openpyxl==3.1.5
playwright==1.48.0
apscheduler==3.10.4
aiofiles==24.1.0
Pillow==10.4.0
jinja2==3.1.4
